--- src/aioway/fn/modes/__init__.py
@@ -1,5 +1,0 @@
-# Copyright (c) AIoWay Authors - All Rights Reserved
-
-from .modules import *
-from .tensors import *
-from .toggles import *
