+++ src/aioway/fn/modules.py
@@ -12,8 +12,8 @@ from collections import abc as cabc
 import torch
 from torch import nn
 
-from ..fn import Thunk, TorchThunk
 from ._on_off import OnOffCtx, OnOffStack
+from .fn import Thunk, TorchThunk
 
 __all__ = [
     "NnFwdFn",
