+++ tests/fn/test_history.py
@@ -5,8 +5,7 @@ import torch
 from torch import nn
 
 from aioway.fake import enabled_fake_mode, torch_fake_mode
-from aioway.fn import fake_fn, module_init, track_fn
-from aioway.fn.modes.modules import module_fwd
+from aioway.fn import fake_fn, module_fwd, module_init, track_fn
 
 
 @pytest.fixture
